pyproject.toml
[project]
name = "reddit-llm-pipeline"
version = "1.0.0"
description = "Real-time Reddit LLM enrichment pipeline - fetches posts, enriches with LLM analysis"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.33.1",
    "pyyaml>=6.0",
    "kafka-python>=2.0.2",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingestion", "src/processing", "src/shared_utils"]
